feat: add record deletion after file deletion, add scheduler
Add the functionality to delete an entry in the database after
deleting a file from the file storage.
Add a new queue to RabbitMQ for the scheduler to
delete old archived files. Add a scheduler.
Add a listener to a new queue via annotations in edo-file-storage.
Add spring-cloud-starter-openfeign dependency to edo-file-storage.